backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
celery==5.3.6
redis==5.0.1
locust==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2
reportlab==4.0.4
